+++ src/common/enums.ts
@@ -19,9 +19,16 @@ export enum PDAStatus {
   'Valid' = 'Valid',
 }
 
+export enum PDAStatusV3 {
+  'EXPIRED' = 'EXPIRED',
+  'PENDING' = 'PENDING',
+  'REVOKED' = 'REVOKED',
+  'SUSPENDED' = 'SUSPENDED',
+  'VALID' = 'VALID',
+}
+
 export enum OrganizationIdentifierType {
   ORG_DID = 'ORG_DID',
-  ORG_ID = 'ORG_ID',
   USERNAME = 'USERNAME',
 }
 
@@ -29,6 +36,12 @@ export enum OrganizationRole {
   Admin = 'Admin',
   Member = 'Member',
   Owner = 'Owner',
+}
+
+export enum OrganizationRoleV3 {
+  ADMIN = 'ADMIN',
+  MEMBER = 'MEMBER',
+  OWNER = 'OWNER',
 }
 
 export enum AuthType {
@@ -42,3 +55,10 @@ export enum Chain {
   EVM = 'EVM',
   SOL = 'SOL',
 }
+
+export type EncryptedAESCipher = {
+  aesBlob: string;
+  keyBlobs: Record<string, string>;
+  iv: string;
+  tag: string;
+};