+++ src/modules/pda/pda.ts
@@ -1,3 +1,4 @@
+import axios from 'axios';
 import {
   CreatePDAInput,
   FilterPDAInput,
@@ -8,9 +9,10 @@ import {
   Sdk,
   UpdatePDAInput,
 } from '../../../gatewaySdk/sources/Gateway';
-import { Chain, SignCipherEnum } from '../../common/enums';
+import { Chain, PDAStatus, SignCipherEnum } from '../../common/enums';
 import { errorHandler, getChain } from '../../helpers/helper';
 import { ValidationService } from '../../services/validator-service';
+import { Config } from '../../common/types';
 
 // secp256k1=evm by default
 // Ed25519=solana
@@ -18,10 +20,16 @@ import { ValidationService } from '../../services/validator-service';
 export class PDA {
   private sdk: Sdk;
   private validationService: ValidationService;
+  private url: string;
+  private apiKey: string;
+  private authToken: string;
 
-  constructor(sdk: Sdk, validationService: ValidationService) {
+  constructor(sdk: Sdk, validationService: ValidationService, config: Config) {
     this.sdk = sdk;
     this.validationService = validationService;
+    this.url = config.url;
+    this.apiKey = config.apiKey;
+    this.authToken = config.token;
   }
 
   /**
@@ -31,7 +39,7 @@ export class PDA {
    * want to query.
    * @returns The function `getPda` is returning a Promise that resolves to a `PDA_queryQuery` object.
    */
-  async getPDA(id: string) {
+  async getPDA(id: number) {
     try {
       this.validationService.validateUUID(id);
       return await this.sdk.PDA_query({ id });
@@ -121,8 +129,8 @@ export class PDA {
   async changePDAStatus(input: UpdatePDAStatusInput) {
     try {
       const chain: Chain = getChain(input.signingCipher as SignCipherEnum);
-      this.validationService.validateObjectProperties(input.data);
       this.validationService.validateWalletAddress(input.signingKey, chain);
+      this.validationService.validateObjectProperties(input.data);
       return await this.sdk.changePDAStatus_mutation({ input });
     } catch (error) {
       throw new Error(errorHandler(error));
@@ -138,10 +146,11 @@ export class PDA {
   async createPDA(pdaInput: CreatePDAInput) {
     try {
       const chain: Chain = getChain(pdaInput.signingCipher as SignCipherEnum);
-      this.validationService.validateObjectProperties(pdaInput.data);
       this.validationService.validateWalletAddress(pdaInput.signingKey, chain);
+
+      this.validationService.validateObjectProperties(pdaInput.data);
       return await this.sdk.createPDA_mutation({ input: pdaInput });
-    } catch (error) {
+    } catch (error: any) {
       throw new Error(errorHandler(error));
     }
   }
@@ -150,19 +159,62 @@ export class PDA {
    * The function `updatePDA` updates a PDA  using the provided input and returns
    * the result of the mutation.
    * @param {UpdatePDAInput} updatedPDA - The parameter `updatedPDA` is of type `UpdatePDAInput`. It is
-   * an input object that contains the data to update a PDA. The specific
-   * properties and their types within `UpdatePDAInput` would depend on the implementation of the
-   * `updatePDA_m
+   * an input object that contains the data to update a PDA.
    * @returns a Promise that resolves to an object of type `updatePDA_mutationMutation`.
    */
   async updatePDA(updatedPDA: UpdatePDAInput) {
     try {
-      this.validationService.validateObjectProperties(updatedPDA.data);
       this.validationService.validateDID(updatedPDA.did);
       this.validationService.validateString(updatedPDA.signature);
+
+      this.validationService.validateObjectProperties(updatedPDA.data);
       return await this.sdk.updatePDA_mutation({ input: updatedPDA });
     } catch (error) {
       throw new Error(errorHandler(error));
+    }
+  }
+
+  /**
+   * The function `uploadFileAsPDA` takes a file path as input and a PDA Id
+   * and updates the PDA and makes it Valid
+   * @param {File} file - The parameter `file` is of type `File`. Maximum file size allowed is 30 MB
+   * @param {number} pdaId - The parameter `pdaId` is of type `number`. It should be valid pdaId
+   * @returns a promise of type fetch response.
+   */
+  async uploadFileAsPDA(
+    file: Buffer,
+    pdaId: number,
+    fileName: string,
+    fileType: string,
+  ) {
+    try {
+      const { PDA: filePda } = await this.getPDA(pdaId);
+
+      if (filePda === undefined || filePda === null)
+        throw new Error(`${pdaId} not found!`);
+
+      if (filePda.status !== PDAStatus.PENDING)
+        throw new Error(
+          `${pdaId} should be in Pending status only. To upload a file`,
+        );
+
+      const formData = new FormData();
+      formData.append('pdaId', BigInt(pdaId).toString());
+      formData.append('file', new Blob([file], { type: fileType }), fileName);
+
+      return await axios.post(
+        `${this.url.replace('/graphql', '')}/file/upload`,
+        formData,
+        {
+          headers: {
+            'x-api-key': this.apiKey,
+            Authorization: `Bearer ${this.authToken}`,
+            'Content-Type': 'multipart/form-data',
+          },
+        },
+      );
+    } catch (error) {
+      throw new Error('File Upload failed!');
     }
   }
 }