+++ test/stubs/v3/user.stub.ts
@@ -5,7 +5,6 @@ export const userStub = (overrideUser?: Partial<User>): User => ({
   createdAt: new Date('2021-01-01T12:00:00Z'),
   updatedAt: new Date('2021-01-01T12:00:00Z'),
   username: 'test01',
-  isCompleted: true,
   issuedPDAs: [],
   receivedPDAs: [],
   receivedProofs: [],
@@ -15,10 +14,11 @@ export const userStub = (overrideUser?: Partial<User>): User => ({
   did: 'did:gatewayid:abc123',
   encryptionKey: '',
   usernameLastUpdated: new Date('2021-01-01T12:00:00Z'),
+  totalFileSize: 0,
   ...overrideUser,
 });
 
-export const activitiesStub = (): Activity[] => {
+export const activitiesStub = (): any => {
   return [
     {
       action: 'PDA_ISSUANCE',
@@ -29,12 +29,10 @@ export const activitiesStub = (): Activity[] => {
         creator: '',
         dataModel: '',
         issuer: '',
-        dataModels: [''],
         organization: '',
         owner: '',
-        pda: '',
+        pda: 1,
         proof: '',
-        requestTemplate: '',
         signedBy: '',
         status: '',
       },